feat: turn off navbar elements not accessible to unauthorized users
